Add files via upload
You will need to install gimp to view these images unfortunately, I wanted to keep the layers separated for ease of manipulation and keep the layers apart. This is the general Idea I had for the card fronts, if we like these then I will go ahead and create a full deck.